fix: wizard spell and persistence flow
This patch tightens the wizard system around combat, persistence, and admin tooling. It adds player-targeted wizard debug commands, clamps wizard skill values against progression config, allows broader integer skill assignment, and preserves wizard data across DB round-trips.

It also fixes spell casting validation by rejecting offensive casts from protection zones, honoring infinite mana, and checking targetability with the correct area-combat rules. Wizard area resolution now respects caster direction for LINE/CONE/WAVE patterns, and spell impact logic tracks actual affected creatures instead of assuming a full area hit.

The change set includes related Docker/local-compose updates and new integration/unit coverage for wizard persistence, area resolution, and spell-caster behavior.

+++ src/creatures/players/player.cpp
@@ -6761,7 +6761,7 @@ const WizardSkillSnapshot &Player::getWizardSkills() const {
 	return wizardSkills;
 }
 
-void Player::setWizardSkill(const WizardSkill skill, const uint16_t value) {
+void Player::setWizardSkill(const WizardSkill skill, const int32_t value) {
 	wizardSkills.set(skill, value);
 }
